changed the weight, float to double

+++ clustering_cliqueness.hpp
@@ -30,10 +30,10 @@ using namespace std;
 struct Node{
 	uint index;
 	uint degree;
-	float CC;
+	double CC;
 	vector<vector<uint>> cluster;
 	vector<uint> neighbors;
-	unordered_map<uint, float> neighbToWeight;
+	unordered_map<uint, double> neighbToWeight;
 	bool operator <(const Node&n) const
     {
         return (degree < n.degree);
@@ -43,11 +43,11 @@ struct Node{
 
 vector<uint> removeDuplicates(vector<uint>& vec);
 
-vector<float> removeDuplicatesCC(vector<float>& vec);
+vector<double> removeDuplicatesCC(vector<double>& vec);
 
 bool findBridge(vector<Node>& vecNodes, set<uint>& cluster, set<uint>& toRemove);
 
-void DFS(uint n, vector<Node>& vecNodes, unordered_set<uint>& visited, set<uint>& nodesInConnexComp, bool& above, float cutoff);
+void DFS(uint n, vector<Node>& vecNodes, unordered_set<uint>& visited, set<uint>& nodesInConnexComp, bool& above, double cutoff);
 
 vector<string> split(const string &s, char delim);
 
@@ -55,32 +55,32 @@ vector<string> split(const string &s, char delim);
 void parsingSRC(ifstream & refFile, vector<Node>& vecNodes);
 
 
-float getCC(unordered_set<uint>& neighbors, vector<Node>& vecNodes);
+double getCC(unordered_set<uint>& neighbors, vector<Node>& vecNodes);
 
 
-int getDeltaCC(set<uint>& toRemove, set<uint>& clust1, vector<Node>& vecNodes, float cutoff);
+int getDeltaCC(set<uint>& toRemove, set<uint>& clust1, vector<Node>& vecNodes, double cutoff);
 
-void computeCCandDeg(vector<Node>& vecNodes, vector<float>& ClCo);
+void computeCCandDeg(vector<Node>& vecNodes, vector<double>& ClCo);
 
 void sortVecNodes(vector<Node>& vecNodes);
 
-void computePseudoCliques(vector<float>& cutoffs, vector<Node>& vecNodes, uint nbThreads, vector<uint>& nodesInOrderOfCC);
+void computePseudoCliques(vector<double>& cutoffs, vector<Node>& vecNodes, uint nbThreads, vector<uint>& nodesInOrderOfCC);
 
-float computeUnionCC(set<uint>& unionC, vector<Node>& vecNodes);
+double computeUnionCC(set<uint>& unionC, vector<Node>& vecNodes);
 
 void transfer(uint tf, uint te, set<uint>& toFill, set<uint>& toEmpty, vector<Node>& vecNodes, vector<set<uint>>& clusters, uint ind);
 
 void merge(uint i1, uint i2, set<uint>& clust1, set<uint>& clust2, vector<set<uint>>& clusters,  vector<Node>& vecNodes, uint ind);
 
-vector<set<uint>> assignNewClusters(set<uint>& clust, vector<Node>& vecNodes, float cutoff);
+vector<set<uint>> assignNewClusters(set<uint>& clust, vector<Node>& vecNodes, double cutoff);
 
 void removeSplittedElements(uint index, vector<set<uint>>& clusters, set<uint>& interC);
 
 //~ uint splitClust(uint i1, uint i2, set<uint>& clust1, set<uint>& clust2, vector<set<uint>>& clusters,  vector<Node>& vecNodes, set<uint>& interC, uint cutoff, uint ind);
-float splitClust(uint i1, uint i2, set<uint>& clust1, set<uint>& clust2, vector<set<uint>>& clusters,  vector<Node>& vecNodes, set<uint>& interC, uint cutoff, uint ind);
+double splitClust(uint i1, uint i2, set<uint>& clust1, set<uint>& clust2, vector<set<uint>>& clusters,  vector<Node>& vecNodes, set<uint>& interC, uint cutoff, uint ind);
 
 //~ uint computeClustersAndCut(float cutoff, vector<Node>& vecNodes, vector<set<uint>>& clusters, uint ind, uint prevCut, vector<uint>& nodesInOrderOfCC);
-float computeClustersAndCut(float cutoff, vector<Node>& vecNodes, vector<set<uint>>& clusters, uint ind, float prevCut, vector<uint>& nodesInOrderOfCC);
+double computeClustersAndCut(double cutoff, vector<Node>& vecNodes, vector<set<uint>>& clusters, uint ind, double prevCut, vector<uint>& nodesInOrderOfCC);
 
 void getVecNodes(vector<Node>& vecNodes, vector<Node>& vecNodesGlobal, set<uint>& nodesInConnexComp);
 
@@ -88,6 +88,6 @@ void cutBrigdesInConnectedComp(vector<Node>& vecNodes, uint val);
 
 bool findArticulPoint(set<uint>& cluster, vector<Node>& vecNodes, set<uint>& interC);
 
-void preProcessGraph(vector<Node>& vecNodes, float cutoff);
+void preProcessGraph(vector<Node>& vecNodes, double cutoff);
 
 #endif
